fix: 聊天上传图片 image_url 400 - OmniAgentExecutor 也走 VLM 描述
- 将 VLM 描述抽取到 AgentImageAttachmentSupport 共享方法（内部，含缓存/限流/重试）
- AgentOrchestrator 改用共享方法，删除私有 VLM 字段和方法
- OmniAgentExecutor.buildCurrentUserMessage 也走 VLM 描述
  → 用户上传图片先 VLM 描述转文本，失败才降级 image_url
- buildInitialMessages/buildCurrentUserMessage 改为 suspend
